+++ src-tauri/src/lib.rs
@@ -1167,6 +1167,7 @@ pub fn run() {
         .plugin(tauri_plugin_opener::init())
         .plugin(tauri_plugin_dialog::init())
         .plugin(tauri_plugin_clipboard_manager::init())
+        .plugin(tauri_plugin_screenshots::init())
         // NOTE: window-state plugin disabled - causes startup hangs on first launch
         // .plugin(tauri_plugin_window_state::Builder::default().build())
         .plugin(tauri_plugin_fs::init())